% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mmpp.r
\docType{package}
\name{mmpp}
\alias{mmpp}
\alias{mmpp-package}
\alias{package-mmpp}
\title{A package for Computing Similarity and Distance Metrics for Marked Point Process Instances}
\description{
A package for Computing Similarity and Distance Metrics for Marked Point Process Instances
}

